+++ CHANGELOG.md
@@ -109,7 +109,7 @@ If needed, pluralize to `Tasks`, `PRs` or `Authors` and list multiple entries se
 ### Added
 - Made `AutoCorrection` expressible by Dictionary literals and updated the `README.md` accordingly.  
   Issue: [#5](https://github.com/Flinesoft/AnyLint/issues/5) | PR: [#11](https://github.com/Flinesoft/AnyLint/pull/11) | Author: [Cihat Gündüz](https://github.com/Jeehut)
-- Added option to skip checks within file contents by specifying `AnyLint.skipHere: <CheckInfo.ID>` or `AnyLint.skipInFile: <All or CheckInfo.ID>`. Checkout the [Skip file content checks](https://github.com/Flinesoft/AnyLint#skip-file-content-checks) README section for more info.  
+- Added option to skip checks within file contents by specifying `AnyLint.skipHere: <Check.ID>` or `AnyLint.skipInFile: <All or Check.ID>`. Checkout the [Skip file content checks](https://github.com/Flinesoft/AnyLint#skip-file-content-checks) README section for more info.  
   Issue: [#9](https://github.com/Flinesoft/AnyLint/issues/9) | PR: [#12](https://github.com/Flinesoft/AnyLint/pull/12) | Author: [Cihat Gündüz](https://github.com/Jeehut)
 
 ## [0.2.0] - 2020-04-10
@@ -122,7 +122,7 @@ If needed, pluralize to `Tasks`, `PRs` or `Authors` and list multiple entries se
 - Added two simple lint check examples in first code sample in README. (Thanks for the pointer, [Dave Verwer](https://github.com/daveverwer)!)  
   Author: [Cihat Gündüz](https://github.com/Jeehut)
 ### Changed
-- Changed `CheckInfo` id casing convention from snake_case to UpperCamelCase in `blank` template.  
+- Changed `Check` id casing convention from snake_case to UpperCamelCase in `blank` template.  
   Author: [Cihat Gündüz](https://github.com/Jeehut)
 
 ## [0.1.0] - 2020-03-22